Refine tracer store naming and docs
