+++ connections_delegate.go
