+++ trace_delegate.go
